requirements.txt
pathway
kafka-python>=2.0.2
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
confluent-kafka>=2.3.0
pydantic>=2.5.0
httpx>=0.26.0
numpy>=1.26.0
# Satellite data (all use stdlib urllib — no extra deps needed)
# Open-Meteo: free, no auth, no extra package
# WAQI: free token, no extra package
# NASA FIRMS: free MAP_KEY, no extra package
